added main target statements
