--- CODE/Basic jaba codes/table.java
@@ -1,11 +1,0 @@
-import java.util.*;
-public class table {
-    public static void main(String[] args) {
-        Scanner sc=new Scanner(System.in);
-        System.out.println("enter a number to print table for ");
-        int n=sc.nextInt();
-        for(int i=1;i<=10;i++){
-            System.out.println(n+"*"+i+"="+(n*i));
-        }
-    }
-}
