+++ CODE/Basic java codes/table.java
@@ -1,0 +1,12 @@
+// Multiplication table of user input number from 1 to 10
+import java.util.*;
+public class table {
+    public static void main(String[] args) {
+        Scanner sc=new Scanner(System.in);
+        System.out.println("enter a number to print table for ");
+        int n=sc.nextInt();
+        for(int i=1;i<=10;i++){
+            System.out.println(n+"*"+i+"="+(n*i));
+        }
+    }
+}
